# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "audio-pipeline",
    "summary": "Recording sessions and their audio segments. Audio is uploaded over the streaming websocket described in docs/streaming-protocol.md.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/sessions/{session_id}/segments": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "List a session's segments",
        "description": "In capture order (by `sequence`), interleaved across resources.",
        "operationId": "list_session_segments_v1_sessions__session_id__segments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "A recording session belonging to you.",
              "title": "Session Id"
            },
            "description": "A recording session belonging to you."
          },
          {
            "name": "resource",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only segments of this resource.",
              "title": "Resource"
            },
            "description": "Only segments of this resource."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SegmentRead_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/segments": {
      "get": {
        "tags": [
          "segments"
        ],
        "summary": "List your segments",
        "description": "Across every session the caller owns, most recently ingested first.",
        "operationId": "list_segments_v1_segments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only segments of this resource.",
              "title": "Resource"
            },
            "description": "Only segments of this resource."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SegmentRead_"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/segments/{segment_id}": {
      "get": {
        "tags": [
          "segments"
        ],
        "summary": "Fetch one segment's metadata",
        "operationId": "get_segment_v1_segments__segment_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "A segment belonging to you.",
              "title": "Segment Id"
            },
            "description": "A segment belonging to you."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/segments/{segment_id}/media": {
      "get": {
        "tags": [
          "segments"
        ],
        "summary": "Download a segment's stored bytes",
        "description": "Serve what the segment stores: the blob for blob-backed resources\n(honouring `Range` and conditional requests), the inline payload as JSON\notherwise.\n\nA `Range` is passed through to the object store, so only that slice is\ntransferred. Segments are immutable, so `If-None-Match` answers 304 without\ntouching the store.",
        "operationId": "get_segment_media_v1_segments__segment_id__media_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "A segment belonging to you.",
              "title": "Segment Id"
            },
            "description": "A segment belonging to you."
          }
        ],
        "responses": {
          "200": {
            "description": "The stored object (blob resources) or inline payload.",
            "content": {
              "audio/*": {},
              "application/json": {}
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "206": {
            "content": {
              "audio/*": {}
            },
            "description": "The requested byte range."
          },
          "304": {
            "description": "The client's cached copy is still current."
          },
          "416": {
            "description": "The requested range lies outside the object."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioAttrs": {
        "properties": {
          "codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codec"
          },
          "sample_rate_hz": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Rate Hz"
          },
          "channels": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          }
        },
        "type": "object",
        "title": "AudioAttrs",
        "description": "PCM parameters as declared by the client (all optional)."
      },
...
      "AudioSegmentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "ingested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ingested At"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "offset_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset Ms"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "byte_size": {
            "type": "integer",
            "title": "Byte Size"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "checksum_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum Sha256",
            "description": "Hex-encoded SHA-256 of the wire payload."
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "resource": {
            "type": "string",
            "const": "audio",
            "title": "Resource",
            "default": "audio"
          },
          "attrs": {
            "$ref": "#/components/schemas/AudioAttrs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "sequence",
          "ingested_at",
          "captured_at",
          "offset_ms",
          "duration_ms",
          "byte_size",
          "content_type",
          "checksum_sha256",
          "metadata",
          "attrs"
        ],
        "title": "AudioSegmentRead"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "The body of every non-2xx response."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationSegmentAttrs": {
        "properties": {
          "points": {
            "type": "integer",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "points"
        ],
        "title": "LocationSegmentAttrs"
      },
      "LocationSegmentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "ingested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ingested At"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "offset_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset Ms"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "byte_size": {
            "type": "integer",
            "title": "Byte Size"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "checksum_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum Sha256",
            "description": "Hex-encoded SHA-256 of the wire payload."
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "resource": {
            "type": "string",
            "const": "location",
            "title": "Resource",
            "default": "location"
          },
          "attrs": {
            "$ref": "#/components/schemas/LocationSegmentAttrs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "sequence",
          "ingested_at",
          "captured_at",
          "offset_ms",
          "duration_ms",
          "byte_size",
          "content_type",
          "checksum_sha256",
          "metadata",
          "attrs"
        ],
        "title": "LocationSegmentRead"
      },
...
      "Page_SegmentRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SegmentRead"
            },
            "type": "array",
            "title": "Items"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "True when more rows exist past this slice."
          }
        },
        "type": "object",
        "required": [
          "items",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "Page[SegmentRead]"
      },
...
      "SegmentRead": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/AudioSegmentRead"
          },
          {
            "$ref": "#/components/schemas/LocationSegmentRead"
          }
        ],
        "discriminator": {
          "propertyName": "resource",
          "mapping": {
            "audio": "#/components/schemas/AudioSegmentRead",
            "location": "#/components/schemas/LocationSegmentRead"
          }
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "An API token from `tokens issue`.",
        "scheme": "bearer"
      }
    }
  }
}